requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain-community
flask
pypdf
pinecone-client
python-dotenv
-e .